+++ src/components/index.js
@@ -7,3 +7,4 @@ export Author from './Author'
 export PageHeader from './PageHeader'
 export StarCount from './StarCount'
 export GithubRepositoryLink from './GithubRepositoryLink'
+export Link from './Link'
