+++ src/pages/Loading.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import Page from './Page'
-import styles from './styles'
+import styles from '../styles'
 
 export default class Loading extends Component {
   render() {
