+++ src/pages/TodoList2.js
@@ -3,7 +3,7 @@ import { Link } from 'react-router'
 
 import { PageHeader } from '../components'
 import Page from './Page'
-import styles from './styles'
+import styles from '../styles'
 
 const containerStyle = {
   display: 'flex',
